Add and change error messages

+++ server/controllers/work.ts
@@ -146,6 +146,13 @@ export const deleteOne = [
     async function (req: Request, res: Response, next: NextFunction) {
         try {
             validationResult(req).throw();
+
+            const instance = await Work.findById(req.params.id);
+
+            if (!instance) {
+                return res.status(404).json({ error: "This work does not exist." });
+            }
+
             const result = await Work.findByIdAndRemove(req.params.id);
             return res.json({ acknowledged: true, deleted: result });
         } catch (error) {
